PropertiesService set default propertyStatus

+++ src/main/java/edu/codegym/toam/controller/AccountRestController.java
@@ -8,6 +8,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.NotActiveException;
+
 @RestController
 @RequestMapping("api/account")
 @CrossOrigin(origins = "*")
@@ -28,7 +30,6 @@ public class AccountRestController {
     public ResponseEntity<Iterable<Account>> getHost() {
         return ResponseEntity.ok(this.accountService.findAllHost());
     }
-
 
     //Lấy danh sách tất cả những thằng thuê nhà
     @PreAuthorize("hasRole('ROLE_ADMIN')")
@@ -67,36 +68,32 @@ public class AccountRestController {
         }
     }
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> removeAccountById(@PathVariable Long id) {
-        try {
-            this.accountService.removeById(id);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
 
-//    Chuyển đổi trạng thái account (từ block sang không block và ngược lại)
+    //    Chuyển đổi trạng thái account (từ block sang không block và ngược lại)
     @PostMapping("/edit/{accountId}")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN','ROLE_RENTER','ROLE_HOST')")
     public ResponseEntity<Account> changeAccountStatus(@PathVariable Long accountId) {
         try {
-           this.accountService.changeAccountStatus(accountId);
+            this.accountService.changeAccountStatus(accountId);
             return new ResponseEntity<>(HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         }
     }
 
-//    @DeleteMapping("/{id}")
-//    public ResponseEntity<String> removePropertyById(@PathVariable Long id) {
-//        try {
-//            this.propertiesService.removeById(id);
-//            return new ResponseEntity<>(HttpStatus.OK);
-//        } catch (Exception e) {
-//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//        }
-//    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> removePropertyById(@PathVariable Long id) {
+        try {
+            if (!this.accountService.checkAccountConstraint(id)) {
+                this.accountService.removeById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else throw new NotActiveException();
+        } catch (NotActiveException nae) {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
 }
